quiz/docker-requirements.txt
astroid==2.2.5
beautifulsoup4==4.8.1
cffi==1.13.2
colorama==0.4.1
Django==2.1
django-cors-headers==3.1.1
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
isort==4.3.19
lazy-object-proxy==1.4.1
mccabe==0.6.1
numpy==1.17.4
pycparser==2.19
Pygments==2.4.0
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.9
pylint-plugin-utils==0.5
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
soupsieve==1.9.4
sqlparse==0.3.0
typed-ast==1.3.5
waitress==1.3.1
WebOb==1.8.5
wrapt==1.11.1
